adds error handling and logging to ProductImport. adjusts db schema with minor field type changes
Image.variant_ids change to bigint array type, unused fields removed from variant table and product table

+++ db/migrate/20190814191911_create_images.rb
@@ -4,7 +4,7 @@ class CreateImages < ActiveRecord::Migration[6.0]
       t.references :product, null: false, foreign_key: true
       t.datetime :created_at
       t.integer :position
-      t.integer :variant_ids, array: true #variant ids associated with the image.
+      t.bigint :variant_ids, array: true #variant ids associated with the image.
       t.string :src
       t.integer :width
       t.integer :height
